aegis-firewall/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis-firewall"
version = "1.0.0"
description = "Security firewall for AI agent tool execution with MCP protocol support"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "AEGIS Team"}
]
keywords = ["security", "firewall", "mcp", "ai-agents", "llm-security"]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pyyaml>=6.0",
    "pynacl>=1.5.0",
    "websockets>=12.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
aegis = "aegis.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
